[qa] Fix Travis CI build errors #93
Travis CI builds for Debian 9, Debian 10, and
Fedora 28 have been fixed. Python3 is set as
the ansible_python_interpreter for Fedora 28
because it drops support for python-firewall
which is for Python2. Cron is also added to
resolve errors in Debian 10. Additionally,
cryptography from pip is installed to fix an
odd SSL error when installing django-redis on
Debian 9.

Fixes #93